+++ plugin/internal/js/transpilation.go
@@ -1,4 +1,4 @@
-package jsutil
+package js
 
 import (
 	"errors"
